Criado Exception e algumas validacoes

+++ Tech-Challenge/src/main/java/com/fiap/techChallenge/TechChallenge/controller/form/EletrodomesticoResultForm.java
@@ -20,5 +20,5 @@ public class EletrodomesticoResultForm {
     private String modelo;
 
     @JsonProperty
-    private Integer potencia;
+    private Double potencia;
 }
